error handling added to lab18

+++ Lab18/handlers/delete.js
@@ -8,49 +8,75 @@ module.exports = (req, res) => {
     if(path == 'faculties') {
         Faculty.findOne({where:{faculty:code}})
         .then(result => {
-            Faculty.destroy({where:{faculty:code}})
-            .then(result2 => {
-                res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
-                res.end(JSON.stringify(result))})
-            .catch(err => {res.writeHead(500); res.end(JSON.stringify(err))})
+            if(result != null) {
+                Faculty.destroy({where:{faculty:code}})
+                .then(result2 => {
+                    res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
+                    res.end(JSON.stringify(result))})
+                .catch(err => {res.writeHead(500); res.end(JSON.stringify({err: err.message}))})
+            } else {
+                res.writeHead(500, {'Content-Type': 'application/json; charset=utf-8'})
+                res.end(JSON.stringify({err: 'faculty not found'}))
+            }
+            
         })
         
     } else if(path == 'pulpits') {
         Pulpit.findOne({where:{pulpit:code}})
         .then(result => {
-            Pulpit.destroy({where:{pulpit:code}})
-            .then(result2 => {
-                res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
-                res.end(JSON.stringify(result))})
-            .catch(err => {res.writeHead(500); res.end(JSON.stringify(err))})
+            if(result != null) {
+                Pulpit.destroy({where:{pulpit:code}})
+                .then(result2 => {
+                    res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
+                    res.end(JSON.stringify(result))})
+                .catch(err => {res.writeHead(500); res.end(JSON.stringify({err: err.message}))})
+            } else {
+                res.writeHead(500, {'Content-Type': 'application/json; charset=utf-8'})
+                res.end(JSON.stringify({err: 'pulpit not found'}))
+            }
         })
         
     } else if(path == 'subjects') {
         Subject.findOne({where:{subject:code}})
         .then(result => {
-            Subject.destroy({where:{subject:code}})
-            .then(result2 => {
-                res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
-                res.end(JSON.stringify(result))})
-            .catch(err => {res.writeHead(500); res.end(JSON.stringify(err))})
+            if(result != null) {
+                Subject.destroy({where:{subject:code}})
+                .then(result2 => {
+                    res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
+                    res.end(JSON.stringify(result))})
+                .catch(err => {res.writeHead(500); res.end(JSON.stringify({err: err.message}))})
+            } else {
+                res.writeHead(500, {'Content-Type': 'application/json; charset=utf-8'})
+                res.end(JSON.stringify({err: 'subject not found'}))
+            }
         })
     } else if(path == 'auditoriumtypes') {
         Auditorium_type.findOne({where:{auditorium_type:code}})
         .then(result => {
-            Auditorium_type.destroy({where:{auditorium_type:code}})
-            .then(result2 => {
-                res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
-                res.end(JSON.stringify(result))})
-            .catch(err => {res.writeHead(500); res.end(JSON.stringify(err))})
+            if(result != null) {
+                Auditorium_type.destroy({where:{auditorium_type:code}})
+                .then(result2 => {
+                    res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
+                    res.end(JSON.stringify(result))})
+                .catch(err => {res.writeHead(500); res.end(JSON.stringify({err: err.message}))})
+            } else {
+                res.writeHead(500, {'Content-Type': 'application/json; charset=utf-8'})
+                res.end(JSON.stringify({err: 'auditorium type not found'}))
+            }
         })
     } else if(path == 'auditoriums') {
         Auditorium.findOne({where:{auditorium:code}})
         .then(result => {
-            Auditorium.destroy({where:{auditorium:code}})
-            .then(result2 => {
-                res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
-                res.end(JSON.stringify(result))})
-            .catch(err => {res.writeHead(500); res.end(JSON.stringify(err))})
+            if(result != null) {
+                Auditorium.destroy({where:{auditorium:code}})
+                .then(result2 => {
+                    res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'})
+                    res.end(JSON.stringify(result))})
+                .catch(err => {res.writeHead(500); res.end(JSON.stringify({err: err.message}))})
+            } else {
+                res.writeHead(500, {'Content-Type': 'application/json; charset=utf-8'})
+                res.end(JSON.stringify({err: 'auditorium not found'}))
+            }
         })
     }
 }
